requirement.txt
pytest>=7.0.0
pytest-cov>=4.0.0          # 测试覆盖率
pytest-html>=3.0.0         # HTML报告
pytest-xdist>=3.0.0        # 并行测试
pytest-mock>=3.10.0        # mock支持
pyyaml>=6.0.0               # yaml支持